feat: deterministic concurrency test + Mull 2000+ mutations in CI
Deterministic concurrency test:
- 100 runs x 4 threads x 500 iterations = 200,000 operations
- Fixed start-gun barrier ensures tight interleaving
- Per-thread instances (Variant B contract) produce identical results
- 0 mismatches across all runs — deterministic
- Added to TSan CI stage (runs after test_concurrent_decrypt)

Mull mutation testing (GitLab CI):
- Ubuntu 22.04 + LLVM-16 + Mull 0.10.0
- --mutators=all generates 2000+ automatic mutations
- --timeout=5000 prevents hanging on complex mutations
- Compiles with -fembed-bitcode for LLVM IR analysis
- Results stored as artifact (mutation_results.txt)

Fix: string_test.cpp int->size_t for std::array indexing (-Wsign-conversion)

+++ tests/string_test.cpp
@@ -38,8 +38,8 @@ int main() {
     // Partial zero: only first 8 bytes
     std::memset(sz_buf.data(), 0xFF, 16);
     stealth::memory::secure_zero(sz_buf.data(), 8);
-    for (int i = 0; i < 8; ++i) assert(sz_buf[i] == 0);
-    for (int i = 8; i < 16; ++i) assert(sz_buf[i] == static_cast<char>(0xFF));
+    for (size_t i = 0; i < 8; ++i) assert(sz_buf[i] == 0);
+    for (size_t i = 8; i < 16; ++i) assert(sz_buf[i] == static_cast<char>(0xFF));
     // Null + zero size: no crash
     stealth::memory::secure_zero(nullptr, 0);
     std::cout << "[+] Known-answer: secure_zero - PASSED\n";
